Extracted Exceptions (Calculator)
