+++ resources/krow/zeale/guis/calculator/exceptions/EmptyEquationException.java
@@ -1,0 +1,10 @@
+package krow.zeale.guis.calculator.exceptions;
+
+public class EmptyEquationException extends Exception {
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+
+}
