+++ resources/krow/zeale/guis/calculator/exceptions/UnmatchedParenthesisException.java
@@ -1,0 +1,10 @@
+package krow.zeale.guis.calculator.exceptions;
+
+public class UnmatchedParenthesisException extends Exception {
+
+	/**
+	 * SVUID
+	 */
+	private static final long serialVersionUID = 1L;
+
+}
